Move deprecated `Router` to deprecated component's file.
